+++ server/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][keepachangelog],
 and this project adheres to [Semantic Versioning][semver].
 
+## [Unreleased]
+
+### Added:
+
+- support http URLs in `${buildout:extends}`
+
 ## [0.2.1] - 2020-04-25
 
 ### Fixed